@@ -1,7 +1,7 @@
 {
 	"name": "react-hello",
 	"engines": {
-		"node": "16.x"
+		"node": "20.x"
 	},
 	"version": "1.0.1",
 	"description": "",
@@ -19,32 +19,32 @@
 	"devDependencies": {
 		"@babel/cli": "^7.4.3",
 		"@babel/core": "^7.4.3",
+		"@babel/eslint-parser": "^7.25.8",
 		"@babel/plugin-proposal-class-properties": "^7.12.1",
 		"@babel/preset-env": "^7.4.3",
 		"@babel/preset-react": "^7.0.0",
 		"@babel/runtime": "^7.12.5",
-		"babel-eslint": "^10.0.1",
 		"babel-loader": "^8.0.5",
-		"babel-plugin-transform-class-properties": "^6.24.1",
 		"bc-console": "0.0.2",
-		"css-loader": "^2.1.1",
-		"dotenv-webpack": "^7.0.3",
-		"eslint": "^4.19.1",
-		"eslint-loader": "^2.0.0",
-		"eslint-plugin-react": "^7.7.0",
-		"file-loader": "^1.1.5",
-		"gh-pages": "^1.2.0",
-		"html-loader": "^0.5.5",
-		"html-webpack-plugin": "^3.2.0",
-		"parse-github-url": "^1.0.2",
-		"prettier": "^1.14.3",
-		"prettier-webpack-plugin": "^1.1.0",
-		"remote-origin-url": "^1.0.0",
-		"style-loader": "^0.19.0",
-		"webpack": "^4.41.2",
-		"webpack-cli": "^3.3.0",
-		"webpack-dev-server": "^3.11.2",
-		"webpack-merge": "^5.8.0"
+		"css-loader": "^7.1.2",
+		"dotenv-webpack": "^8.1.0",
+		"eslint": "^9.12.0",
+		"eslint-plugin-react": "^7.37.1",
+		"eslint-webpack-plugin": "^4.2.0",
+		"file-loader": "^6.2.0",
+		"gh-pages": "^6.2.0",
+		"html-loader": "^5.1.0",
+		"html-webpack-plugin": "^5.6.1",
+		"parse-github-url": "^1.0.3",
+		"prettier": "^3.3.3",
+		"remote-origin-url": "^2.0.0",
+		"sass": "^1.80.1",
+		"sass-loader": "^16.0.2",
+		"style-loader": "^4.0.0",
+		"webpack": "^5.95.0",
+		"webpack-cli": "^5.1.4",
+		"webpack-dev-server": "^5.1.0",
+		"webpack-merge": "^6.0.1"
 	},
 	"babel": {
 		"presets": [
@@ -63,25 +63,24 @@
 	},
 	"dependencies": {
 		"14": "^3.1.6",
-		"@babel/plugin-transform-runtime": "^7.12.1",
-		"@sweetalert/with-react": "^0.1.1",
-		"@uiw/react-md-editor": "^2.1.1",
+		"@babel/plugin-transform-runtime": "^7.25.7",
+		"@uiw/react-md-editor": "^4.0.4",
 		"bc-react-notifier": "^1.1.6",
-		"bootstrap": "^4.1.3",
-		"jquery": "^3.5.1",
-		"popper.js": "^1.12.9",
-		"prop-types": "^15.6.1",
-		"query-string": "^5.1.1",
-		"react": "^16.8.6",
-		"react-dnd": "^10.0.2",
-		"react-dnd-html5-backend": "^10.0.2",
-		"react-dom": "^16.8.6",
-		"react-lightweight-tooltip": "^1.0.0",
-		"react-multi-select-component": "^4.2.1",
+		"bootstrap": "^5.3.3",
+		"jquery": "^3.7.1",
+		"popper.js": "^1.16.1",
+		"prop-types": "^15.8.1",
+		"query-string": "^9.1.1",
+		"react": "^18.3.1",
+		"react-dnd": "^16.0.1",
+		"react-dnd-html5-backend": "^16.0.1",
+		"react-dom": "^18.3.1",
+		"react-multi-select-component": "^4.3.4",
 		"react-polyfills": "0.0.1",
-		"react-scroll": "^1.8.1",
-		"react-select": "^4.3.1",
-		"react-toast-notifications": "^2.4.4",
-		"sweetalert": "^2.1.2"
+		"react-scroll": "^1.9.0",
+		"react-select": "^5.8.1",
+		"react-toastify": "^10.0.6",
+		"react-tooltip": "^5.28.0",
+		"sweetalert2": "^11.14.3"
 	}
 }
