@@ -2,6 +2,7 @@ const { merge } = require('webpack-merge');
 const common = require('./webpack.common.js');
 const webpack = require('webpack');
 const Dotenv = require('dotenv-webpack');
+const ESLintPlugin = require('eslint-webpack-plugin');
 const path = require('path');
 
 module.exports = merge(common, {
@@ -13,13 +14,18 @@ module.exports = merge(common, {
   //   publicPath: '/'
   // },
   devServer: {
-    contentBase:  './dist',
+    static: {
+      directory: './dist'
+    },
     hot: true,
-    disableHostCheck: true,
+    allowedHosts: 'all',
     historyApiFallback: true
   },
   plugins: [
     new webpack.HotModuleReplacementPlugin(),
+    new ESLintPlugin({
+      extensions: ['js', 'jsx'],
+    }),
     new Dotenv({
       safe: true,
       systemvars: true
